Avgerage aggs should output always double values (ESQL-582)
This change addresses two specific things:

1. Update the result type of average aggregators to output double
values. Otherwise 1 + 2 / 2 will equal 1. This aligns grouping and non
grouping. And was only an issue for non-grouping long avg.

2. Update the aggregate mapper to use the field data type. The type
specific aggs are driven by the type of values they accumulate (not
their output type. The output type can be inferred but the input type)

+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/aggregation/AvgLongAggregator.java
@@ -13,7 +13,7 @@ import org.elasticsearch.compute.ann.Aggregator;
 import org.elasticsearch.compute.ann.GroupingAggregator;
 import org.elasticsearch.compute.data.Block;
 import org.elasticsearch.compute.data.BlockBuilder;
-import org.elasticsearch.compute.data.LongVector;
+import org.elasticsearch.compute.data.DoubleVector;
 import org.elasticsearch.core.Releasables;
 
 import java.lang.invoke.MethodHandles;
@@ -60,11 +60,11 @@ class AvgLongAggregator {
 
     public static Block evaluateFinal(GroupingAvgState state) {
         int positions = state.largestGroupId + 1;
-        long[] result = new long[positions];
+        double[] result = new double[positions];
         for (int i = 0; i < positions; i++) {
-            result[i] = state.values.get(i) / state.counts.get(i);
+            result[i] = (double) state.values.get(i) / state.counts.get(i);
         }
-        return new LongVector(result, positions).asBlock();
+        return new DoubleVector(result, positions).asBlock();
     }
 
     static class AvgState implements AggregatorState<AvgLongAggregator.AvgState> {
